refactor: siginInFormSlice if문을 삼항연산자로 정리
