+++ src/lib/store/index.ts
@@ -3,8 +3,8 @@ import logger from 'redux-logger'
 import storage from 'redux-persist/lib/storage'
 import persistReducer from 'redux-persist/es/persistReducer'
 import persistStore from 'redux-persist/es/persistStore'
-import userFormSlice from './signForm/userFormSlice'
-import signInFormSlice from './signForm/signInFormSlice'
+import userFormSlice from './member/userFormSlice'
+import signInFormSlice from './auth/signInFormSlice'
 import memberSlice from './member/memberSlice'
 import editModeSlice from './member/editModeSlice'
 import userMenuSlice from './userMenuSlice'
